database connections and repo controller stuff
