+++ Pet_Pillbox/Data/UsersRepo.cs
@@ -5,14 +5,18 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
 
 namespace Pet_Pillbox.Data
 {
     public class UsersRepository
     {
-        static List<User> _users = new List<User>();
+        readonly string _connectionString;
 
-        const string _connectionString = "Server=localhost;Database=Pet_Pillbox;Trusted_Connection=True;";
+        public UsersRepository(IConfiguration configuration)
+        {
+            _connectionString = configuration.GetConnectionString("Pet_Pillbox");
+        }
 
         public List<User> GetAllUsers()
         {
