+++ Pet_Pillbox/Pet_Pillbox.csproj
@@ -6,9 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Controllers\" />
     <Folder Include="Models\" />
-    <Folder Include="Data\" />
   </ItemGroup>
 
   <ItemGroup>
