+++ Pet_Pillbox/appsettings.json
@@ -6,5 +6,8 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "ConnectionStrings": {
+    "Pet_Pillbox": "Server=localhost;Database=Pet_Pillbox;Trusted_Connection=True;"
+  },
   "AllowedHosts": "*"
 }
